Update main.cc
There was a memory leak where `raft` wasn't getting deleted. It is preferred to now use smart pointers to take care of scope management. In the changed code, we create a shared pointer to a `Raft` object. If we successfully assign a value to `raft`, we go ahead and de-reference it below.